+++ packages/php-wasm/progress/src/lib/custom-event.ts
@@ -1,0 +1,17 @@
+export const CustomEvent =
+	globalThis.CustomEvent ??
+	class extends globalThis.Event {
+		detail = null;
+		constructor(
+			name: string,
+			options: {
+				detail?: any;
+				bubbles?: boolean;
+				cancellable?: boolean;
+				composed?: boolean;
+			} = {}
+		) {
+			super(name, options);
+			this.detail = options.detail;
+		}
+	};
